requirements.txt
beautifulsoup4==4.12.3
requests==2.32.3
pandas==2.2.3
boto3==1.35.81
python-dotenv==1.0.1